Moved CodeExample into its own file.
